third_party/blink/web_tests/platform/generic/svg/dom/svg-root-lengths-expected.txt

This tests the behavior of root SVG length value resolution

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Initial/default values:
PASS svg.width.baseVal.value is 200
PASS svg.height.baseVal.value is 200


Updated relative values:
PASS svg.width.baseVal.value is 100
PASS svg.height.baseVal.value is 20


Updated fixed values:
PASS svg.width.baseVal.value is 150
PASS svg.height.baseVal.value is 50


viewBox has no effect on top level length resolution.
PASS svg.width.baseVal.value is 200
PASS svg.height.baseVal.value is 100
PASS successfullyParsed is true

TEST COMPLETE
